requirements.txt
torch==2.0.1+cu117
torchvision==0.15.2+cu117
torchaudio==2.0.2+cu117
-f https://download.pytorch.org/whl/torch_stable.html
transformers==4.37.2
tiktoken==0.6.0
verovio==4.3.1
accelerate==0.28.0
opencv-python==4.10.0.84
flask==3.0.3
Flask_Cors==5.0.0
langchain_core==0.3.8
langchain_groq==0.2.0
python-dotenv